refactor: move serializeJsx into componentToImage and simplify
- Moved serializeJsx from standalone file into componentToImage.ts as private function
- Removed dead code paths that were never covered in tests:
  - Function component type handling
  - Complex React element types with name properties
  - Function props serialization
  - Try/catch fallback for non-React objects
- Simplified implementation to only handle OG image use cases
- Improved test coverage by eliminating uncovered branches
- Added AIDEV-NOTE documenting the function's purpose and scope
